Remove two tests that cannot run using PHPUnit 6
They result in SimpleSAMLphp attempting to set the session name, but
that cannot be done because headers have already been sent (because
PHPUnit has already begun writing results to the console). I don't know
how this used to work with older versions of PHPUnit.

Given what these are testing, the risk of removing these tests seems less
than the risk of using an unsupported version of PHP.